requirements.txt
Django==3.0.11
djangorestframework==3.12.2
slack-sdk==3.0.0